easysurv: Simplify Survival Data Analysis and Model Fitting

Inspect survival data, plot Kaplan-Meier curves, assess the proportional hazards assumption, fit parametric survival models, predict and plot survival and hazards, and export the outputs to 'Excel'. A simple interface for fitting survival models using flexsurv::flexsurvreg(), flexsurv::flexsurvspline(), flexsurvcure::flexsurvcure(), and survival::survreg().

See also

Author

Maintainer: Niall Davison niall.davison@maplehealthgroup.com (ORCID)

Authors:

Brad Kievit brad.kievit@maplehealthgroup.com

Other contributors:

Maple Health Group, LLC [copyright holder, funder]