Fit parametric mixture cure survival models for health economic evaluations

fit.models.cure fits parametric mixture cure survival models for health economic evaluations.The code for fit.models.cure is primarily taken from the equivalent DARTH function, with minor edits to functionality and documentation.

Usage

fit.models.cure(
  formula = NULL,
  data,
  distr = NULL,
  method = "mle",
  weights = NULL,
  ...
)

Arguments

formula

A formula specifying the model to be used, in the form Surv(time,event)~strata[+covariates] for /codeflexsurv. Recommend that class(strata) == "factor".

data

A data frame containing the data to be used for the analysis. This must contain data for the 'event' variable.

distr

A (vector of) string(s) containing the name(s) of the model(s) to be fitted. Available options are: "exponential","gamma", "genf", "gengamma","gompertz","weibull", "weibullPH","loglogistic","lognormal".

method

The method used to do the estimation. Current possible values are "mle".

weights

Optional for case weights. The function expects a string corresponding to a variable name within the data.

...

Additional arguments for runMLE.cure.

Value

A list containing the fitted parametric mixture cure survival models for each specified distribution. The list includes the model objects and additional information such as AIC, BIC, and KM data. This object is analogous to the object generated from the fit.models function, but with an added cure fraction.

Examples

if (FALSE) {
library(survival)

fit <- fit.models.cure(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull"), method = "mle"
)
print(fit)
}