Modified version of ggcoxdiagnostics to replace gather_ and set geom_smooth formulae

Displays diagnostics graphs presenting goodness of Cox Proportional Hazards Model fit, that can be calculated with coxph function. This function is largely identical to ggcoxdiagnostics from the /codesurvminer package with some minor alterations in order to accommodate easysurv functionality.

Usage

plot_schoenfeld(
  fit_coxph,
  data,
  formula,
  hline = TRUE,
  sline = TRUE,
  sline.se = TRUE,
  hline.col = "red",
  hline.size = 1,
  hline.alpha = 1,
  hline.yintercept = 0,
  hline.lty = "dashed",
  sline.col = "blue",
  sline.size = 1,
  sline.alpha = 0.3,
  sline.lty = "dashed",
  point.col = "black",
  point.size = 1,
  point.shape = 19,
  point.alpha = 1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  font.family = "Roboto Condensed",
  plot.theme = theme_easysurv()
)

Arguments

fit_coxph

An object of class coxph.object - created with coxph function.

data

Data for fit_coxph

formula

Formula for fit_coxph

hline

a logical - should the horizontal line be added to highlight the Y=0 level.

sline, sline.se

a logical - should the smooth line be added to highlight the local average for residuals.

hline.col, hline.size, hline.lty, hline.alpha, hline.yintercept

color, size, linetype, visibility and Y-axis coordinate to be used for geom_hline. Used only when hline = TRUE.

sline.col, sline.size, sline.lty, sline.alpha

color, size, linetype and visibility to be used for geom_smooth. Used only when sline = TRUE.

point.col, point.size, point.shape, point.alpha

color, size, shape and visibility to be used for points.

title, subtitle, caption

main title, subtitle and caption.

font.family

Font for the plot. Default is Roboto Condensed.

plot.theme

ggplot2 theme for the plot. Default is theme_easysurv()

Value

Returns an object of class ggplot.

Examples

if (FALSE) {
#to add
}