Predict method for fit_models

Predict method for fit_models

Usage

# S3 method for fit_models
predict(object, eval_time = NULL, type = c("survival", "hazard"), ...)

Arguments

object

An object of class fit_models

eval_time

(Optional) A vector of evaluation time points for generating predictions. Default is NULL, which if left as NULL, generates a sequence from 0 to 5 times the maximum observed time.

type

A character vector indicating the type of predictions to generate. Default is c("survival", "hazard").

...

Additional arguments

Value

A list of predictions for each model in the fit_models object.

Examples

models <- fit_models(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  predict_by = "group",
  covariates = "group"
)

predict(models)
#> $Good
#> $Good$profiles
#>    profile group
#> 1 profile1  Good
#> 
#> $Good$predicted_surv
#> # A tibble: 100 × 8
#>    .eval_time Exponential Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl> <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0           1     1            1        1              1    
#>  2      0.374       0.978 0.995        1.00     0.983          0.996
#>  3      0.747       0.956 0.984        0.998    0.966          0.986
#>  4      1.12        0.935 0.968        0.988    0.948          0.972
#>  5      1.49        0.914 0.950        0.970    0.930          0.955
#>  6      1.87        0.893 0.930        0.946    0.911          0.934
#>  7      2.24        0.873 0.908        0.917    0.891          0.912
#>  8      2.62        0.854 0.885        0.886    0.872          0.887
#>  9      2.99        0.835 0.860        0.855    0.851          0.862
#> 10      3.36        0.816 0.835        0.824    0.830          0.835
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Good$predicted_hazards
#> # A tibble: 100 × 8
#>    .eval_time Exponential  Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl>  <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0          0.0604 0        NaN          0.0440         0     
#>  2      0.374      0.0604 0.0234     0.000756   0.0463         0.0196
#>  3      0.747      0.0604 0.0364     0.0142     0.0486         0.0327
#>  4      1.12       0.0604 0.0463     0.0380     0.0510         0.0437
#>  5      1.49       0.0604 0.0544     0.0598     0.0536         0.0534
#>  6      1.87       0.0604 0.0613     0.0760     0.0563         0.0618
#>  7      2.24       0.0604 0.0672     0.0871     0.0591         0.0692
#>  8      2.62       0.0604 0.0724     0.0942     0.0621         0.0757
#>  9      2.99       0.0604 0.0770     0.0986     0.0652         0.0813
#> 10      3.36       0.0604 0.0811     0.101      0.0685         0.0861
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Good$bshazard
#>           time     hazard         lcl        ucl
#> 1   0.02191781 0.01169274 0.003930704 0.03478260
#> 2   0.04657534 0.01200839 0.004115436 0.03503915
#> 3   0.12602740 0.01308461 0.004764573 0.03593336
#> 4   0.15616438 0.01351762 0.005033385 0.03630280
#> 5   0.31232877 0.01599927 0.006642671 0.03853522
#> 6   0.40547945 0.01768783 0.007796044 0.04013051
#> 7   0.53424658 0.02031097 0.009652168 0.04274020
#> 8   0.56164384 0.02091090 0.010088943 0.04334108
#> 9   0.58356164 0.02140357 0.010449727 0.04383970
#> 10  0.62739726 0.02242402 0.011201806 0.04488887
#> 11  0.74794521 0.02548813 0.013478492 0.04819863
#> 12  0.81095890 0.02720203 0.014794148 0.05001644
#> 13  0.87397260 0.02903095 0.016201130 0.05202083
#> 14  0.88219178 0.02927840 0.016391222 0.05229779
#> 15  1.00821918 0.03331335 0.019482285 0.05696349
#> 16  1.16712329 0.03889543 0.023848145 0.06343697
#> 17  1.18356164 0.03952379 0.024327547 0.06421239
#> 18  1.26301370 0.04257715 0.026687163 0.06792831
#> 19  1.30136986 0.04403110 0.027850042 0.06961345
#> 20  1.30410959 0.04413683 0.027934003 0.06973793
#> 21  1.44931507 0.05011962 0.032519477 0.07724529
#> 22  1.46027397 0.05060277 0.032873655 0.07789339
#> 23  1.48493151 0.05163095 0.033657828 0.07920162
#> 24  1.50136986 0.05225786 0.034169117 0.07992258
#> 25  1.50684932 0.05246852 0.034339993 0.08016733
#> 26  1.51506849 0.05278610 0.034596693 0.08053869
#> 27  1.52602740 0.05321253 0.034939626 0.08104190
#> 28  1.53150685 0.05342704 0.035111356 0.08129702
#> 29  1.55068493 0.05418465 0.035713626 0.08220884
#> 30  1.56164384 0.05462238 0.036058506 0.08274342
#> 31  1.63835616 0.05778713 0.038479909 0.08678170
#> 32  1.67671233 0.05943761 0.039689478 0.08901175
#> 33  1.72328767 0.06144509 0.041141366 0.09176892
#> 34  1.74520548 0.06224136 0.041795291 0.09268954
#> 35  1.78356164 0.06365974 0.042934376 0.09438968
#> 36  1.78630137 0.06376228 0.043015407 0.09451562
#> 37  1.81369863 0.06479681 0.043822690 0.09580944
#> 38  1.89589041 0.06800224 0.046201283 0.10009038
#> 39  1.90410959 0.06833137 0.046434726 0.10055355
#> 40  1.93698630 0.06966393 0.047358973 0.10247399
#> 41  1.97808219 0.07117016 0.048473116 0.10449486
#> 42  1.99726027 0.07178068 0.048975109 0.10520582
#> 43  2.01095890 0.07221997 0.049330232 0.10573080
#> 44  2.02739726 0.07275068 0.049752353 0.10638011
#> 45  2.03013699 0.07283951 0.049822263 0.10649042
#> 46  2.04657534 0.07337476 0.050238957 0.10716495
#> 47  2.04931507 0.07346435 0.050307934 0.10727951
#> 48  2.06301370 0.07391395 0.050650735 0.10786164
#> 49  2.06849315 0.07409456 0.050786861 0.10809889
#> 50  2.07671233 0.07436630 0.050989959 0.10845951
#> 51  2.08493151 0.07463903 0.051191723 0.10882590
#> 52  2.08767123 0.07473017 0.051258676 0.10894932
#> 53  2.14794521 0.07676352 0.052691468 0.11183286
#> 54  2.18904110 0.07818152 0.053621341 0.11399100
#> 55  2.19452055 0.07837256 0.053742280 0.11429098
#> 56  2.20821918 0.07877775 0.054041550 0.11483633
#> 57  2.35068493 0.08242611 0.056882134 0.11944108
#> 58  2.36986301 0.08292997 0.057222697 0.12018622
#> 59  2.38356164 0.08329175 0.057459420 0.12073766
#> 60  2.40273973 0.08380090 0.057781546 0.12153693
#> 61  2.43835616 0.08475474 0.058350585 0.12310702
#> 62  2.44383562 0.08489500 0.058436483 0.12333325
#> 63  2.50958904 0.08614380 0.059503904 0.12471039
#> 64  2.56438356 0.08719849 0.060291872 0.12611280
#> 65  2.58904110 0.08767730 0.060614834 0.12682225
#> 66  2.64931507 0.08885883 0.061319199 0.12876705
#> 67  2.65479452 0.08896703 0.061377176 0.12895889
#> 68  2.65753425 0.08902118 0.061405784 0.12905577
#> 69  2.66301370 0.08912958 0.061462239 0.12925141
#> 70  2.69041096 0.08963174 0.061743516 0.13011649
#> 71  2.69315068 0.08966505 0.061776064 0.13014460
#> 72  2.70136986 0.08976505 0.061872210 0.13023235
#> 73  2.71506849 0.08993195 0.062027417 0.13039003
#> 74  2.72602740 0.09006570 0.062147001 0.13052651
#> 75  2.74520548 0.09030024 0.062346344 0.13078769
#> 76  2.98356164 0.09290381 0.063935093 0.13499814
#> 77  2.98630137 0.09292075 0.063949013 0.13501796
#> 78  2.98904110 0.09293769 0.063962653 0.13503839
#> 79  2.99726027 0.09298852 0.064001891 0.13510328
#> 80  3.02739726 0.09317516 0.064124103 0.13538763
#> 81  3.03835616 0.09324312 0.064160083 0.13550917
#> 82  3.05205479 0.09332814 0.064198706 0.13567472
#> 83  3.06027397 0.09337919 0.064218493 0.13578134
#> 84  3.06575342 0.09341324 0.064230275 0.13585545
#> 85  3.08219178 0.09351546 0.064258867 0.13609236
#> 86  3.13972603 0.09387411 0.064279825 0.13709353
#> 87  3.15616438 0.09397683 0.064263525 0.13742858
#> 88  3.18630137 0.09406409 0.064275222 0.13765884
#> 89  3.20273973 0.09406946 0.064295151 0.13763188
#> 90  3.20821918 0.09407124 0.064299508 0.13762779
#> 91  3.21643836 0.09407393 0.064303896 0.13762625
#> 92  3.32054795 0.09410793 0.064137389 0.13808330
#> 93  3.33698630 0.09411330 0.064073963 0.13823576
#> 94  3.33972603 0.09411419 0.064062428 0.13826328
#> 95  3.34794521 0.09411688 0.064026183 0.13834944
#> 96  3.37808219 0.09412672 0.063872457 0.13871143
#> 97  3.40547945 0.09413568 0.063704874 0.13910278
#> 98  3.41369863 0.09412493 0.063659095 0.13917105
#> 99  3.46301370 0.09386261 0.063472243 0.13880381
#> 100 3.50410959 0.09364456 0.063246821 0.13865209
#> 101 3.55068493 0.09339805 0.062916425 0.13864737
#> 102 3.60821918 0.09309444 0.062402566 0.13888171
#> 103 3.64109589 0.09292139 0.062059013 0.13913185
#> 104 3.67123288 0.09267267 0.061786815 0.13899766
#> 105 3.67397260 0.09264440 0.061765106 0.13896172
#> 106 3.67671233 0.09261615 0.061743123 0.13892642
#> 107 3.67945205 0.09258790 0.061720867 0.13889176
#> 108 3.69589041 0.09241861 0.061581625 0.13869722
#> 109 3.69863014 0.09239043 0.061557471 0.13866702
#> 110 3.70410959 0.09233408 0.061508355 0.13860854
#> 111 3.71232877 0.09224963 0.061432671 0.13852555
#> 112 3.71506849 0.09222150 0.061406910 0.13849915
#> 113 3.72054795 0.09216526 0.061354589 0.13844824
#> 114 3.73698630 0.09199674 0.061191291 0.13831053
#> 115 3.75616438 0.09180053 0.060988916 0.13817816
#> 116 3.89041096 0.09043869 0.059237950 0.13807292
#> 117 3.90958904 0.09019843 0.059010122 0.13787052
#> 118 3.92328767 0.09001492 0.058857011 0.13766731
#> 119 3.93150685 0.08990500 0.058762059 0.13755319
#> 120 3.94794521 0.08968555 0.058565300 0.13734240
#> 121 3.99726027 0.08903043 0.057921675 0.13684716
#> 122 4.04109589 0.08845211 0.057285579 0.13657496
#> 123 4.06301370 0.08816436 0.056946272 0.13649629
#> 124 4.07123288 0.08805670 0.056815536 0.13647644
#> 125 4.08219178 0.08791335 0.056638336 0.13645806
#> 126 4.10684932 0.08759167 0.056227928 0.13645000
#> 127 4.11506849 0.08748471 0.056087636 0.13645742
#> 128 4.14794521 0.08703754 0.055560707 0.13634695
#> 129 4.16712329 0.08675154 0.055303791 0.13608162
#> 130 4.32328767 0.08445741 0.052807524 0.13507647
#> 131 4.38356164 0.08358419 0.051697058 0.13513955
#> 132 4.43013699 0.08286964 0.050977399 0.13471414
#> 133 4.44931507 0.08257719 0.050663250 0.13459444
#> 134 4.45205479 0.08253549 0.050617562 0.13457992
#> 135 4.45753425 0.08245217 0.050525589 0.13455281
#> 136 4.46301370 0.08236893 0.050432831 0.13452824
#> 137 4.47123288 0.08224422 0.050292240 0.13449614
#> 138 4.49589041 0.08187124 0.049860322 0.13443355
#> 139 4.50684932 0.08170601 0.049663647 0.13442172
#> 140 4.56438356 0.08084403 0.048587352 0.13451561
#> 141 4.58904110 0.08047740 0.048105571 0.13463330
#> 142 4.60000000 0.08031499 0.047887880 0.13469999
#> 143 4.60273973 0.08027443 0.047833129 0.13471803
#> 144 4.63561644 0.07977167 0.047214314 0.13477945
#> 145 4.66027397 0.07937834 0.046783990 0.13468112
#> 146 4.66575342 0.07929119 0.046686328 0.13466669
#> 147 4.70410959 0.07868386 0.045983207 0.13463937
#> 148 4.71232877 0.07855433 0.045828320 0.13464998
#> 149 4.71506849 0.07851120 0.045776376 0.13465478
#> 150 4.71780822 0.07846809 0.045724277 0.13466021
#> 151 4.77534247 0.07756828 0.044596807 0.13491633
#> 152 4.79726027 0.07722821 0.044152100 0.13508297
#> 153 4.80273973 0.07714343 0.044039753 0.13513039
#> 154 4.81095890 0.07701643 0.043870398 0.13520576
#> 155 4.82191781 0.07684742 0.043643090 0.13531412
#> 156 4.83561644 0.07663668 0.043356659 0.13546204
#> 157 4.84109589 0.07655255 0.043241406 0.13552503
#> 158 4.87945205 0.07595590 0.042469001 0.13584728
#> 159 4.89315068 0.07573854 0.042210508 0.13589807
#> 160 4.95068493 0.07483241 0.041086570 0.13629491
#> 161 4.98082192 0.07436211 0.040476235 0.13661655
#> 162 5.02191781 0.07372555 0.039624254 0.13717499
#> 163 5.04109589 0.07343035 0.039220016 0.13748125
#> 164 5.07671233 0.07288527 0.038460027 0.13812424
#> 165 5.09041096 0.07267670 0.038164983 0.13839656
#> 166 5.10410959 0.07246872 0.037868647 0.13868243
#> 167 5.12054795 0.07221395 0.037541546 0.13890891
#> 168 5.16164384 0.07157898 0.036717451 0.13953991
#> 169 5.21917808 0.07069938 0.035531142 0.14067664
#> 170 5.26575342 0.06999524 0.034550523 0.14180202
#> 171 5.27671233 0.06983058 0.034317886 0.14209239
#> 172 5.29589041 0.06954337 0.033909449 0.14262337
#> 173 5.30958904 0.06933893 0.033616856 0.14302015
#> 174 5.41369863 0.06774843 0.031441845 0.14597904
#> 175 5.43561644 0.06741133 0.030985497 0.14665855
#> 176 5.44931507 0.06720150 0.030699406 0.14710518
#> 177 5.49863014 0.06645149 0.029666300 0.14884905
#> 178 5.50410959 0.06636867 0.029551379 0.14905567
#> 179 5.50684932 0.06632730 0.029493920 0.14915994
#> 180 5.51780822 0.06616208 0.029264109 0.14958328
#> 181 5.52602740 0.06603844 0.029091807 0.14990734
#> 182 5.56164384 0.06550531 0.028346345 0.15137563
#> 183 5.58630137 0.06513874 0.027832092 0.15245193
#> 184 5.61369863 0.06470228 0.027269621 0.15351827
#> 185 5.63287671 0.06439481 0.026876541 0.15428666
#> 186 5.65753425 0.06400164 0.026371444 0.15532748
#> 187 5.73424658 0.06279373 0.024809333 0.15893422
#> 188 5.84109589 0.06114094 0.022684369 0.16479254
#> 189 5.88493151 0.06043816 0.021837545 0.16727024
#> 190 5.90684932 0.06008981 0.021418139 0.16858537
#> 191 5.92054795 0.05987311 0.021157587 0.16943279
#> 192 5.94520548 0.05948501 0.020691940 0.17100701
#> 193 5.95068493 0.05939911 0.020589082 0.17136531
#> 194 5.95890411 0.05927049 0.020435237 0.17190851
#> 195 6.00547945 0.05854689 0.019574114 0.17511591
#> 196 6.01369863 0.05842012 0.019424142 0.17570456
#> 197 6.07397260 0.05749879 0.018344214 0.18022638
#> 198 6.12876712 0.05664852 0.017395424 0.18447695
#> 199 6.22191781 0.05523145 0.015844886 0.19252352
#> 200 6.29041096 0.05421216 0.014761537 0.19909566
#> 201 6.29315068 0.05417178 0.014719268 0.19937007
#> 202 6.35616438 0.05324303 0.013770559 0.20586095
#> 203 6.49315068 0.05126832 0.011850115 0.22180720
#> 204 6.49863014 0.05119088 0.011777478 0.22250145
#> 205 6.57808219 0.05007947 0.010761453 0.23304970
#> 206 6.67945205 0.04869073 0.009560421 0.24797941
#> 207 6.70958904 0.04828534 0.009223541 0.25277429
#> 208 6.72876712 0.04802912 0.009014006 0.25591249
#> 209 6.76986301 0.04748465 0.008577655 0.26286816
#> 210 6.95616438 0.04508905 0.006804284 0.29878560
#> 211 6.98904110 0.04467889 0.006524434 0.30595806
#> 212 7.00273973 0.04450909 0.006410684 0.30902460
#> 213 7.28493151 0.04115009 0.004408642 0.38409333
#> 
#> 
#> $Medium
#> $Medium$profiles
#>    profile  group
#> 1 profile2 Medium
#> 
#> $Medium$predicted_surv
#> # A tibble: 100 × 8
#>    .eval_time Exponential Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl> <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0           1     1            1        1              1    
#>  2      0.374       0.950 0.986        0.998    0.962          0.988
#>  3      0.747       0.903 0.957        0.974    0.923          0.960
#>  4      1.12        0.858 0.920        0.925    0.884          0.921
#>  5      1.49        0.815 0.877        0.866    0.845          0.876
#>  6      1.87        0.774 0.832        0.806    0.806          0.826
#>  7      2.24        0.736 0.785        0.749    0.767          0.775
#>  8      2.62        0.699 0.738        0.697    0.728          0.725
#>  9      2.99        0.664 0.691        0.650    0.689          0.676
#> 10      3.36        0.631 0.646        0.607    0.650          0.629
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Medium$predicted_hazards
#> # A tibble: 100 × 8
#>    .eval_time Exponential  Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl>  <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0           0.137 0          NaN         0.102         0     
#>  2      0.374       0.137 0.0627       0.0239    0.107         0.0579
#>  3      0.747       0.137 0.0944       0.108     0.112         0.0949
#>  4      1.12        0.137 0.117        0.162     0.118         0.124 
#>  5      1.49        0.137 0.135        0.186     0.124         0.146 
#>  6      1.87        0.137 0.149        0.195     0.130         0.163 
#>  7      2.24        0.137 0.161        0.195     0.137         0.176 
#>  8      2.62        0.137 0.170        0.191     0.143         0.184 
#>  9      2.99        0.137 0.179        0.186     0.151         0.190 
#> 10      3.36        0.137 0.186        0.179     0.158         0.193 
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Medium$bshazard
#>           time     hazard        lcl        ucl
#> 1   0.04657534 0.03999115 0.01827684 0.08750378
#> 2   0.17260274 0.04743482 0.02410035 0.09336222
#> 3   0.19452055 0.04886412 0.02525774 0.09453347
#> 4   0.46027397 0.06974419 0.04297021 0.11320054
#> 5   0.46301370 0.06999896 0.04319011 0.11344854
#> 6   0.48493151 0.07207096 0.04497462 0.11549232
#> 7   0.52328767 0.07582896 0.04819635 0.11930428
#> 8   0.61095890 0.08461881 0.05599129 0.12788317
#> 9   0.75616438 0.10147456 0.07033768 0.14639503
#> 10  0.78082192 0.10422704 0.07279853 0.14922383
#> 11  0.84383562 0.11153707 0.07923539 0.15700709
#> 12  0.90136986 0.11865832 0.08525700 0.16514534
#> 13  0.92602740 0.12184781 0.08786036 0.16898280
#> 14  0.95890411 0.12623423 0.09133750 0.17446373
#> 15  0.96712329 0.12735529 0.09220632 0.17590301
#> 16  0.98356164 0.12962736 0.09394186 0.17886864
#> 17  1.01095890 0.13283106 0.09661452 0.18262357
#> 18  1.01369863 0.13311091 0.09686729 0.18291535
#> 19  1.01643836 0.13339135 0.09711997 0.18320900
#> 20  1.01917808 0.13367238 0.09737255 0.18350455
#> 21  1.02465753 0.13423622 0.09787739 0.18410139
#> 22  1.05479452 0.13738014 0.10064395 0.18752546
#> 23  1.07945205 0.14000713 0.10288970 0.19051466
#> 24  1.15068493 0.14788179 0.10923538 0.20020092
#> 25  1.16164384 0.14913198 0.11018750 0.20184093
#> 26  1.17534247 0.15070959 0.11136696 0.20395081
#> 27  1.20000000 0.15359146 0.11345800 0.20792133
#> 28  1.23013699 0.15718870 0.11595357 0.21308776
#> 29  1.24931507 0.15855373 0.11721726 0.21446743
#> 30  1.29589041 0.16173536 0.12010608 0.21779352
#> 31  1.30410959 0.16230341 0.12059566 0.21843570
#> 32  1.31780822 0.16325460 0.12139725 0.21954423
#> 33  1.34520548 0.16517375 0.12294435 0.22190827
#> 34  1.35616438 0.16594771 0.12354148 0.22291009
#> 35  1.36438356 0.16653056 0.12398096 0.22368296
#> 36  1.37534247 0.16731088 0.12455559 0.22474247
#> 37  1.37808219 0.16750653 0.12469720 0.22501257
#> 38  1.41917808 0.17046888 0.12672112 0.22931963
#> 39  1.44931507 0.17267452 0.12808352 0.23278945
#> 40  1.47945205 0.17425241 0.12915676 0.23509339
#> 41  1.49041096 0.17441237 0.12940341 0.23507630
#> 42  1.49315068 0.17445238 0.12946285 0.23507620
#> 43  1.50136986 0.17457248 0.12963575 0.23508599
#> 44  1.50684932 0.17465259 0.12974649 0.23510097
#> 45  1.51232877 0.17473273 0.12985358 0.23512273
#> 46  1.51780822 0.17481292 0.12995699 0.23515130
#> 47  1.54520548 0.17521439 0.13041819 0.23539723
#> 48  1.55342466 0.17533501 0.13053818 0.23550477
#> 49  1.58356164 0.17577799 0.13090463 0.23603368
#> 50  1.60547945 0.17610086 0.13109809 0.23655199
#> 51  1.62739726 0.17642433 0.13122994 0.23718326
#> 52  1.70684932 0.17755644 0.13120507 0.24028257
#> 53  1.74520548 0.17686077 0.13104568 0.23869335
#> 54  1.76986301 0.17641499 0.13084759 0.23785115
#> 55  1.81643836 0.17557603 0.13026773 0.23664297
#> 56  1.83561644 0.17523173 0.12995125 0.23628984
#> 57  1.84931507 0.17498622 0.12969782 0.23608860
#> 58  1.88219178 0.17439839 0.12899835 0.23577665
#> 59  1.95616438 0.17291671 0.12722082 0.23502589
#> 60  1.96438356 0.17267172 0.12710505 0.23457387
#> 61  1.98082192 0.17218280 0.12684996 0.23371641
#> 62  1.99178082 0.17185762 0.12666241 0.23317922
#> 63  2.00821918 0.17137100 0.12635488 0.23242489
#> 64  2.01917808 0.17104735 0.12613244 0.23195616
#> 65  2.02739726 0.17080502 0.12595650 0.23162246
#> 66  2.07671233 0.16935821 0.12473914 0.22993749
#> 67  2.09863014 0.16871912 0.12411120 0.22935999
#> 68  2.10958904 0.16840049 0.12377782 0.22910989
#> 69  2.11506849 0.16824139 0.12360638 0.22899439
#> 70  2.16986301 0.16665870 0.12172456 0.22818011
#> 71  2.17808219 0.16642259 0.12141724 0.22810993
#> 72  2.26027397 0.16397696 0.11991929 0.22422116
#> 73  2.29589041 0.16292763 0.11905794 0.22296214
#> 74  2.33698630 0.16172521 0.11790013 0.22184068
#> 75  2.33972603 0.16164537 0.11781689 0.22177825
#> 76  2.34246575 0.16156556 0.11773292 0.22171734
#> 77  2.34794521 0.16140607 0.11756278 0.22160006
#> 78  2.37260274 0.16069029 0.11676149 0.22114629
#> 79  2.37534247 0.16061096 0.11666893 0.22110325
#> 80  2.40000000 0.15989871 0.11580529 0.22078092
#> 81  2.43287671 0.15895668 0.11496452 0.21978282
#> 82  2.43561644 0.15887866 0.11492259 0.21964723
#> 83  2.44109589 0.15872272 0.11483641 0.21938078
#> 84  2.44657534 0.15856694 0.11474716 0.21912066
#> 85  2.51506849 0.15663251 0.11337397 0.21639662
#> 86  2.55616438 0.15548320 0.11232637 0.21522128
#> 87  2.56438356 0.15525436 0.11209749 0.21502636
#> 88  2.57534247 0.15494975 0.11178253 0.21478691
#> 89  2.58082192 0.15479767 0.11162093 0.21467586
#> 90  2.61917808 0.15373729 0.11041495 0.21405755
#> 91  2.64109589 0.15313462 0.10966963 0.21382593
#> 92  2.66849315 0.15247896 0.10909063 0.21312400
#> 93  2.77534247 0.15024494 0.10744941 0.21008530
#> 94  2.83835616 0.14894281 0.10598379 0.20931466
#> 95  2.90136986 0.14772725 0.10446930 0.20889716
#> 96  2.95068493 0.14706120 0.10410528 0.20774159
#> 97  2.95342466 0.14702429 0.10407845 0.20769085
#> 98  2.98082192 0.14665565 0.10377206 0.20726080
#> 99  2.99452055 0.14647168 0.10359312 0.20709824
#> 100 2.99726027 0.14643491 0.10355529 0.20706990
#> 101 3.00000000 0.14639815 0.10351678 0.20704294
#> 102 3.04931507 0.14573810 0.10270983 0.20679221
#> 103 3.06849315 0.14548221 0.10233929 0.20681278
#> 104 3.12328767 0.14475358 0.10111573 0.20722393
#> 105 3.15068493 0.14444933 0.10092914 0.20673523
#> 106 3.16986301 0.14424397 0.10082265 0.20636557
#> 107 3.18904110 0.14403890 0.10068353 0.20606354
#> 108 3.23835616 0.14351292 0.10017720 0.20559527
#> 109 3.26575342 0.14322154 0.09980510 0.20552467
#> 110 3.26849315 0.14319244 0.09976440 0.20552495
#> 111 3.27397260 0.14313425 0.09968112 0.20552951
#> 112 3.30136986 0.14284363 0.09922763 0.20563128
#> 113 3.30958904 0.14275657 0.09907974 0.20568723
#> 114 3.34794521 0.14235095 0.09831998 0.20610046
#> 115 3.36986301 0.14213863 0.09795724 0.20624702
#> 116 3.37260274 0.14211921 0.09795465 0.20619613
#> 117 3.43287671 0.14169272 0.09773611 0.20541871
#> 118 3.49863014 0.14122892 0.09714780 0.20531197
#> 119 3.50684932 0.14117105 0.09704915 0.20535228
#> 120 3.51506849 0.14111320 0.09694504 0.20540437
#> 121 3.55068493 0.14086281 0.09643211 0.20576479
#> 122 3.57808219 0.14067051 0.09597128 0.20618869
#> 123 3.62465753 0.14044671 0.09557382 0.20638788
#> 124 3.66301370 0.14034705 0.09553840 0.20617148
#> 125 3.67945205 0.14030435 0.09548608 0.20615896
#> 126 3.70136986 0.14024745 0.09538183 0.20621693
#> 127 3.71506849 0.14021190 0.09529678 0.20629634
#> 128 3.80000000 0.13999167 0.09443747 0.20752006
#> 129 3.83835616 0.13989233 0.09387084 0.20847651
#> 130 3.92876712 0.14035737 0.09422447 0.20907723
#> 131 4.00000000 0.14072486 0.09403905 0.21058789
#> 132 4.00821918 0.14076732 0.09399190 0.21082071
#> 133 4.02465753 0.14085229 0.09388201 0.21132235
#> 134 4.03287671 0.14089479 0.09381936 0.21159110
#> 135 4.05753425 0.14102238 0.09360116 0.21246863
#> 136 4.06301370 0.14105074 0.09354661 0.21267808
#> 137 4.11506849 0.14162294 0.09383233 0.21375422
#> 138 4.12328767 0.14172834 0.09390264 0.21391223
#> 139 4.17808219 0.14243305 0.09423142 0.21529097
#> 140 4.18356164 0.14250371 0.09425091 0.21546008
#> 141 4.18630137 0.14253906 0.09425974 0.21554677
#> 142 4.26575342 0.14356788 0.09425561 0.21867914
#> 143 4.33424658 0.14471641 0.09441669 0.22181289
#> 144 4.37808219 0.14580654 0.09508043 0.22359540
#> 145 4.38630137 0.14601185 0.09518737 0.22397362
#> 146 4.39452055 0.14621745 0.09528876 0.22436583
#> 147 4.48493151 0.14849825 0.09603846 0.22961352
#> 148 4.49863014 0.14884691 0.09609433 0.23055891
#> 149 4.53424658 0.14975728 0.09617059 0.23320272
#> 150 4.61369863 0.15268635 0.09765547 0.23872827
#> 151 4.61643836 0.15279548 0.09771085 0.23893417
#> 152 4.63835616 0.15367135 0.09813099 0.24064655
#> 153 4.66027397 0.15455224 0.09851044 0.24247576
#> 154 4.67671233 0.15521622 0.09876835 0.24392504
#> 155 4.69589041 0.15599446 0.09904044 0.24570037
#> 156 4.70410959 0.15632919 0.09914761 0.24648922
#> 157 4.71232877 0.15666464 0.09924913 0.24729494
#> 158 4.73698630 0.15767530 0.09952015 0.24981374
#> 159 4.73972603 0.15778800 0.09954718 0.25010305
#> 160 4.81095890 0.16101078 0.10062478 0.25763506
#> 161 4.83013699 0.16202493 0.10114048 0.25956055
#> 162 4.90136986 0.16584802 0.10276842 0.26764609
#> 163 4.95068493 0.16854749 0.10363074 0.27412961
#> 164 4.96986301 0.16960911 0.10390865 0.27685134
#> 165 4.98630137 0.17052439 0.10412175 0.27927466
#> 166 4.98904110 0.17067742 0.10415504 0.27968670
#> 167 5.00273973 0.17144461 0.10431203 0.28178201
#> 168 5.02191781 0.17252448 0.10450567 0.28481417
#> 169 5.02465753 0.17268644 0.10455795 0.28520651
#> 170 5.03561644 0.17336463 0.10486875 0.28659915
#> 171 5.04383562 0.17387502 0.10509416 0.28767082
#> 172 5.04657534 0.17404548 0.10516783 0.28803325
#> 173 5.05753425 0.17472901 0.10545515 0.28950913
#> 174 5.06027397 0.17490031 0.10552514 0.28988467
#> 175 5.07397260 0.17575934 0.10586407 0.29180199
#> 176 5.09863014 0.17731624 0.10642787 0.29542119
#> 177 5.10136986 0.17749008 0.10648685 0.29583678
#> 178 5.11232877 0.17818713 0.10671548 0.29752622
#> 179 5.11780822 0.17853669 0.10682543 0.29838727
#> 180 5.14520548 0.18029477 0.10733179 0.30285717
#> 181 5.19726027 0.18368296 0.10809848 0.31211752
#> 182 5.21643836 0.18494723 0.10831820 0.31578699
#> 183 5.25479452 0.18750193 0.10866023 0.32354960
#> 184 5.35890411 0.19432247 0.11021546 0.34261274
#> 185 5.36712329 0.19487009 0.11029533 0.34429702
#> 186 5.38356164 0.19596998 0.11043440 0.34775607
#> 187 5.41095890 0.19781693 0.11060622 0.35379147
#> 188 5.42191781 0.19856058 0.11065447 0.35630105
#> 189 5.45205479 0.20062005 0.11072872 0.36348660
#> 190 5.50410959 0.20414370 0.11071065 0.37642856
#> 191 5.50958904 0.20445965 0.11072585 0.37754282
#> 192 5.52054795 0.20509302 0.11074573 0.37981735
#> 193 5.52876712 0.20556934 0.11075152 0.38156365
#> 194 5.55342466 0.20700493 0.11072291 0.38701153
#> 195 5.56164384 0.20748569 0.11069835 0.38889750
#> 196 5.57260274 0.20812844 0.11065418 0.39146689
#> 197 5.61917808 0.21088239 0.11032603 0.40309059
#> 198 5.62191781 0.21104552 0.11029990 0.40381009
#> 199 5.63561644 0.21186305 0.11015833 0.40746764
#> 200 5.82465753 0.22269633 0.10702936 0.46336494
#> 201 5.83013699 0.22299388 0.10690537 0.46514290
#> 202 5.85753425 0.22448764 0.10624499 0.47432540
#> 203 5.87397260 0.22538869 0.10581787 0.48007072
#> 204 5.89863014 0.22674705 0.10513649 0.48902361
#> 205 5.96438356 0.23040952 0.10310529 0.51489641
#> 206 6.04109589 0.23448560 0.10070992 0.54595912
#> 207 6.10136986 0.23771754 0.09853104 0.57352107
#> 208 6.11780822 0.23860668 0.09789517 0.58157262
#> 209 6.12054795 0.23875520 0.09778763 0.58293718
#> 210 6.13424658 0.23949915 0.09724352 0.58985776
#> 211 6.44657534 0.25634018 0.08316198 0.79014819
#> 212 6.49863014 0.25905542 0.08057083 0.83292809
#> 213 6.52054795 0.26020726 0.07945443 0.85215913
#> 214 6.54246575 0.26136423 0.07832640 0.87213583
#> 215 7.02191781 0.28696349 0.05334766 1.54361130
#> 216 7.15616438 0.29435893 0.04683227 1.85015974
#> 
#> 
#> $Poor
#> $Poor$profiles
#>    profile group
#> 1 profile3  Poor
#> 
#> $Poor$predicted_surv
#> # A tibble: 100 × 8
#>    .eval_time Exponential Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl> <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0           1     1            1        1              1    
#>  2      0.374       0.900 0.962        0.978    0.918          0.961
#>  3      0.747       0.811 0.892        0.880    0.839          0.881
#>  4      1.12        0.730 0.808        0.769    0.764          0.784
#>  5      1.49        0.657 0.722        0.671    0.692          0.686
#>  6      1.87        0.592 0.637        0.590    0.624          0.597
#>  7      2.24        0.533 0.558        0.523    0.559          0.518
#>  8      2.62        0.479 0.485        0.468    0.499          0.451
#>  9      2.99        0.432 0.420        0.422    0.442          0.394
#> 10      3.36        0.389 0.361        0.383    0.389          0.345
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Poor$predicted_hazards
#> # A tibble: 100 × 8
#>    .eval_time Exponential Gamma `Gen. Gamma` Gompertz `log-Logistic`
#>         <dbl>       <dbl> <dbl>        <dbl>    <dbl>          <dbl>
#>  1      0           0.281 0          NaN        0.223          0    
#>  2      0.374       0.281 0.165        0.187    0.234          0.181
#>  3      0.747       0.281 0.238        0.344    0.246          0.280
#>  4      1.12        0.281 0.285        0.368    0.258          0.338
#>  5      1.49        0.281 0.319        0.355    0.271          0.368
#>  6      1.87        0.281 0.345        0.333    0.285          0.379
#>  7      2.24        0.281 0.365        0.309    0.299          0.377
#>  8      2.62        0.281 0.381        0.288    0.314          0.368
#>  9      2.99        0.281 0.395        0.268    0.330          0.356
#> 10      3.36        0.281 0.407        0.251    0.347          0.341
#> # ℹ 90 more rows
#> # ℹ 2 more variables: `log-Normal` <dbl>, `Weibull (AFT)` <dbl>
#> 
#> $Poor$bshazard
#>           time     hazard        lcl       ucl
#> 1   0.04109589 0.08806728 0.04790201 0.1619107
#> 2   0.04383562 0.08847271 0.04823992 0.1622602
#> 3   0.04931507 0.08928916 0.04892207 0.1629644
#> 4   0.07945205 0.09391623 0.05282717 0.1669645
#> 5   0.11506849 0.09969469 0.05778530 0.1719993
#> 6   0.17808219 0.11080276 0.06749511 0.1818984
#> 7   0.18356164 0.11182528 0.06839650 0.1828294
#> 8   0.19726027 0.11442307 0.07068960 0.1852131
#> 9   0.26849315 0.12890542 0.08348624 0.1990341
#> 10  0.30958904 0.13770518 0.09147832 0.2072919
#> 11  0.32876712 0.14201488 0.09540174 0.2114031
#> 12  0.43835616 0.16936025 0.12001766 0.2389890
#> 13  0.46849315 0.17776360 0.12734926 0.2481357
#> 14  0.47397260 0.17933568 0.12870270 0.2498882
#> 15  0.47945205 0.18092167 0.13006179 0.2516700
#> 16  0.48493151 0.18252168 0.13142629 0.2534817
#> 17  0.49315068 0.18478670 0.13344180 0.2558878
#> 18  0.49589041 0.18549645 0.13410422 0.2565835
#> 19  0.50410959 0.18764211 0.13610368 0.2586966
#> 20  0.50958904 0.18908631 0.13744670 0.2601273
#> 21  0.53424658 0.19572404 0.14358575 0.2667946
#> 22  0.62191781 0.22126916 0.16647855 0.2940922
#> 23  0.63835616 0.22641765 0.17091061 0.2999518
#> 24  0.65205479 0.23079945 0.17462633 0.3050421
#> 25  0.66027397 0.23346913 0.17686380 0.3081910
#> 26  0.66301370 0.23436586 0.17761078 0.3092569
#> 27  0.67671233 0.23890148 0.18135299 0.3147118
#> 28  0.68219178 0.24074021 0.18285272 0.3169537
#> 29  0.68767123 0.24259309 0.18435373 0.3192309
#> 30  0.74520548 0.26031507 0.19966304 0.3393915
#> 31  0.75342466 0.26265785 0.20181003 0.3418519
#> 32  0.76986301 0.26740685 0.20611441 0.3469259
#> 33  0.78356164 0.27142988 0.20970843 0.3513172
#> 34  0.78904110 0.27305599 0.21114698 0.3531169
#> 35  0.80273973 0.27716401 0.21474377 0.3577281
#> 36  0.83561644 0.28727718 0.22336060 0.3694840
#> 37  0.84109589 0.28899823 0.22479222 0.3715430
#> 38  0.84931507 0.29159915 0.22693613 0.3746872
#> 39  0.86575342 0.29687143 0.23120883 0.3811820
#> 40  0.92054795 0.31514389 0.24525021 0.4049565
#> 41  0.92602740 0.31703189 0.24663344 0.4075247
#> 42  0.93972603 0.32147747 0.25001800 0.4133613
#> 43  0.94246575 0.32213536 0.25065189 0.4140052
#> 44  0.95342466 0.32478043 0.25317931 0.4166309
#> 45  0.97808219 0.33081153 0.25881079 0.4228428
#> 46  0.98082192 0.33148853 0.25943115 0.4235599
#> 47  0.98356164 0.33216691 0.26005034 0.4242827
#> 48  0.98630137 0.33284668 0.26066834 0.4250110
#> 49  1.01369863 0.33972137 0.26677673 0.4326112
#> 50  1.02739726 0.34321178 0.26977714 0.4366357
#> 51  1.03287671 0.34461797 0.27096633 0.4382889
#> 52  1.03835616 0.34602991 0.27214898 0.4399675
#> 53  1.07397260 0.35534964 0.27966551 0.4515157
#> 54  1.10410959 0.36343135 0.28577475 0.4621904
#> 55  1.12328767 0.36866969 0.28953383 0.4694351
#> 56  1.13698630 0.37245753 0.29215549 0.4748314
#> 57  1.14246575 0.37398353 0.29318916 0.4770425
#> 58  1.15068493 0.37628428 0.29472348 0.4804159
#> 59  1.16712329 0.38018810 0.29768771 0.4855524
#> 60  1.19452055 0.38482264 0.30233590 0.4898143
#> 61  1.22191781 0.38951367 0.30675143 0.4946054
#> 62  1.22739726 0.39045872 0.30760465 0.4956297
#> 63  1.24657534 0.39378448 0.31050859 0.4993943
#> 64  1.26027397 0.39617735 0.31250200 0.5022575
#> 65  1.26849315 0.39762004 0.31366482 0.5040466
#> 66  1.27397260 0.39858476 0.31442596 0.5052694
#> 67  1.29041096 0.40149297 0.31664091 0.5090833
#> 68  1.33150685 0.40885667 0.32171961 0.5195946
#> 69  1.33698630 0.40984865 0.32234670 0.5211033
#> 70  1.34246575 0.41084303 0.32296197 0.5226374
#> 71  1.34520548 0.41134113 0.32326518 0.5234140
#> 72  1.36986301 0.41585126 0.32586182 0.5306920
#> 73  1.38082192 0.41787160 0.32694004 0.5340939
#> 74  1.41095890 0.42044239 0.32968514 0.5361837
#> 75  1.43835616 0.42239082 0.33185595 0.5376249
#> 76  1.44109589 0.42258616 0.33205521 0.5377993
#> 77  1.45205479 0.42336842 0.33281928 0.5385530
#> 78  1.46575342 0.42434828 0.33369931 0.5396219
#> 79  1.46849315 0.42454452 0.33386523 0.5398527
#> 80  1.47123288 0.42474086 0.33402776 0.5400892
#> 81  1.48219178 0.42552711 0.33464398 0.5410924
#> 82  1.49041096 0.42611775 0.33507040 0.5419050
#> 83  1.49589041 0.42651196 0.33533762 0.5424755
#> 84  1.49863014 0.42670921 0.33546611 0.5427694
#> 85  1.50136986 0.42690654 0.33559117 0.5430691
#> 86  1.54246575 0.42987756 0.33705844 0.5482572
#> 87  1.56438356 0.43147055 0.33753142 0.5515541
#> 88  1.56986301 0.43186971 0.33761672 0.5524355
#> 89  1.57534247 0.43226925 0.33768902 0.5533396
#> 90  1.58082192 0.43266916 0.33774844 0.5542664
#> 91  1.58630137 0.43306944 0.33779507 0.5552157
#> 92  1.62739726 0.43418336 0.33799450 0.5577463
#> 93  1.63013699 0.43410810 0.33800139 0.5575416
#> 94  1.63287671 0.43403285 0.33800486 0.5573426
#> 95  1.64383562 0.43373199 0.33798441 0.5566039
#> 96  1.68493151 0.43260561 0.33741425 0.5546524
#> 97  1.70410959 0.43208097 0.33688088 0.5541839
#> 98  1.70684932 0.43200607 0.33679088 0.5541398
#> 99  1.70958904 0.43193119 0.33669745 0.5541015
#> 100 1.72054795 0.43163178 0.33628955 0.5540047
#> 101 1.72328767 0.43155696 0.33617907 0.5539947
#> 102 1.72876712 0.43140736 0.33594797 0.5539915
#> 103 1.73150685 0.43133258 0.33582738 0.5539983
#> 104 1.77534247 0.43013786 0.33345129 0.5548594
#> 105 1.78082192 0.42998876 0.33309689 0.5550647
#> 106 1.80000000 0.42946729 0.33176069 0.5559494
#> 107 1.84931507 0.42702330 0.32829075 0.5554494
#> 108 1.86027397 0.42600824 0.32765061 0.5538919
#> 109 1.91232877 0.42121960 0.32390422 0.5477729
#> 110 1.95068493 0.41772560 0.32041903 0.5445827
#> 111 1.97534247 0.41549477 0.31787226 0.5430984
#> 112 1.97808219 0.41524764 0.31757517 0.5429600
#> 113 2.00000000 0.41327586 0.31510111 0.5420385
#> 114 2.00273973 0.41303005 0.31478000 0.5419462
#> 115 2.00547945 0.41278438 0.31445633 0.5418589
#> 116 2.02739726 0.41082430 0.31177745 0.5413368
#> 117 2.03287671 0.41033574 0.31108372 0.5412544
#> 118 2.04109589 0.40960398 0.31002589 0.5411658
#> 119 2.04657534 0.40911687 0.30930947 0.5411299
#> 120 2.05753425 0.40806486 0.30787882 0.5408522
#> 121 2.06575342 0.40674449 0.30697002 0.5389487
#> 122 2.10410959 0.40063903 0.30240656 0.5307809
#> 123 2.10684932 0.40020645 0.30206067 0.5302418
#> 124 2.11506849 0.39891151 0.30100747 0.5286593
#> 125 2.12602740 0.39719143 0.29956740 0.5266295
#> 126 2.13424658 0.39590624 0.29846099 0.5251666
#> 127 2.16438356 0.39122937 0.29421807 0.5202278
#> 128 2.18356164 0.38828198 0.29137350 0.5174214
#> 129 2.18630137 0.38786274 0.29095844 0.5170412
#> 130 2.20547945 0.38494072 0.28799504 0.5145205
#> 131 2.24383562 0.37916254 0.28178732 0.5101870
#> 132 2.26575342 0.37589975 0.27809120 0.5081089
#> 133 2.26849315 0.37549388 0.27762237 0.5078685
#> 134 2.29041096 0.37165697 0.27409146 0.5039519
#> 135 2.30410959 0.36883521 0.27202552 0.5000980
#> 136 2.30684932 0.36827343 0.27160641 0.4993451
#> 137 2.34246575 0.36104773 0.26598541 0.4900850
#> 138 2.34794521 0.35994874 0.26509327 0.4887453
#> 139 2.35342466 0.35885309 0.26419415 0.4874277
#> 140 2.35890411 0.35776078 0.26328819 0.4861319
#> 141 2.37534247 0.35450375 0.26053053 0.4823731
#> 142 2.41917808 0.34596258 0.25290975 0.4732522
#> 143 2.58904110 0.31316923 0.22458224 0.4366996
#> 144 2.61917808 0.30740607 0.21977966 0.4299692
#> 145 2.62739726 0.30585278 0.21844718 0.4282313
#> 146 2.63013699 0.30533677 0.21800101 0.4276611
#> 147 2.66575342 0.29870727 0.21211774 0.4206439
#> 148 2.66849315 0.29820331 0.21165936 0.4201336
#> 149 2.68767123 0.29469932 0.20843060 0.4166744
#> 150 2.85753425 0.26679705 0.18554713 0.3836258
#> 151 2.90958904 0.25891717 0.17851357 0.3755350
#> 152 2.95890411 0.25184991 0.17205176 0.3686587
#> 153 2.98356164 0.24895824 0.17000510 0.3645785
#> 154 2.98630137 0.24863900 0.16977173 0.3641440
#> 155 2.99452055 0.24768373 0.16906462 0.3628626
#> 156 3.00000000 0.24704893 0.16858746 0.3620268
#> 157 3.01369863 0.24546903 0.16737485 0.3600006
#> 158 3.03561644 0.24296217 0.16537814 0.3569433
#> 159 3.15616438 0.22962515 0.15335001 0.3438390
#> 160 3.18356164 0.22688862 0.15082436 0.3413139
#> 161 3.19726027 0.22578524 0.15006042 0.3397230
#> 162 3.20547945 0.22512579 0.14958858 0.3388068
#> 163 3.22465753 0.22359455 0.14844916 0.3367788
#> 164 3.24109589 0.22229035 0.14743070 0.3351609
#> 165 3.24657534 0.22185731 0.14708286 0.3346458
#> 166 3.30684932 0.21714919 0.14299754 0.3297523
#> 167 3.35616438 0.21337150 0.13933764 0.3267415
#> 168 3.37534247 0.21192021 0.13784784 0.3257953
#> 169 3.39726027 0.21027368 0.13610502 0.3248596
#> 170 3.42465753 0.20888425 0.13509197 0.3229846
#> 171 3.64657534 0.19917340 0.12553502 0.3160078
#> 172 3.70410959 0.19821099 0.12455120 0.3154333
#> 173 3.71780822 0.19798253 0.12425211 0.3154641
#> 174 3.73424658 0.19770872 0.12386127 0.3155848
#> 175 3.74246575 0.19757196 0.12365300 0.3156792
#> 176 3.80273973 0.19657195 0.12187503 0.3170504
#> 177 3.94794521 0.19663247 0.12007525 0.3220008
#> 178 3.95342466 0.19667668 0.12002676 0.3222758
#> 179 3.96986301 0.19680936 0.11985861 0.3231635
#> 180 3.99726027 0.19703068 0.11950402 0.3248518
#> 181 4.03287671 0.19731878 0.11890875 0.3274334
#> 182 4.09041096 0.19821834 0.11838372 0.3318912
#> 183 4.11506849 0.19899452 0.11869880 0.3336076
#> 184 4.27397260 0.20406999 0.11889345 0.3502679
#> 185 4.30136986 0.20508002 0.11883492 0.3539180
#> 186 4.34794521 0.20761341 0.11990651 0.3594745
#> 187 4.35342466 0.20791351 0.12001456 0.3601898
#> 188 4.39178082 0.21002636 0.12066489 0.3655668
#> 189 4.50684932 0.21649461 0.12153005 0.3856652
#> 190 4.52876712 0.21792547 0.12184290 0.3897766
#> 191 4.56438356 0.22067729 0.12296392 0.3960387
#> 192 4.58904110 0.22260272 0.12364711 0.4007532
#> 193 4.66027397 0.22825997 0.12519499 0.4161717
#> 194 4.66301370 0.22848040 0.12524204 0.4168193
#> 195 4.71780822 0.23293407 0.12599389 0.4306422
#> 196 4.73698630 0.23451328 0.12617395 0.4358782
#> 197 4.73972603 0.23473975 0.12619624 0.4366433
#> 198 4.75342466 0.23600268 0.12661241 0.4399037
#> 199 4.79726027 0.24025137 0.12818247 0.4503012
#> 200 4.85205479 0.24566995 0.12980412 0.4649600
#> 201 4.90684932 0.25121074 0.13105063 0.4815455
#> 202 4.94794521 0.25544819 0.13174610 0.4952995
#> 203 4.96986301 0.25775850 0.13210889 0.5029142
#> 204 5.00273973 0.26152903 0.13344476 0.5125524
#> 205 5.03561644 0.26535472 0.13464002 0.5229732
#> 206 5.07945205 0.27054284 0.13601481 0.5381284
#> 207 5.08219178 0.27087044 0.13609247 0.5391246
#> 208 5.08493151 0.27119845 0.13616917 0.5401267
#> 209 5.41643836 0.31702047 0.14567906 0.6898862
#> 210 5.42739726 0.31887320 0.14621050 0.6954365
#> 211 5.51780822 0.33457721 0.14990967 0.7467291
#> 212 5.54520548 0.33948698 0.15079137 0.7643104
#> 213 5.61095890 0.35156640 0.15246782 0.8106559
#> 214 5.64109589 0.35727493 0.15309312 0.8337760
#> 215 6.26301370 0.52118585 0.16561287 1.6401787
#> 216 6.72876712 0.70756410 0.14975815 3.3430364
#> 217 6.75890411 0.72183393 0.14786656 3.5237462
#> 
#>