Generate a data frame of predicted survival from a fit.models or fit.models.cure object

This function generates a data frame of predicted survival values based on a fit.models or fit.models.cure object. It takes a set of times and the group in the data, and returns a data frame with predicted survival values for each model in the input object at the specified times.

Usage

predict_fits(fits, t, group = 1)

Arguments

fits

An object obtained from fit.models or fit.models.cure.

t

A vector of times at which to calculate predicted survival, e.g., seq(from = 0, to = 1000, length.out = 100)

group

In cases of multiple stratas within one fit.models object, the index of the "strata" of interest (e.g., In data frames with two treatments, group = "1" or group = "2")

Value

A data frame with predicted survival values for each model at the specified times.

Examples

if (FALSE) {
library(survHE)

# Single-arm
fits <- fit.models(Surv(time, status) ~ 1,
  data = lung,
  dist = c("exponential", "weibull")
)
t <- seq(from = 0, to = 1000, length.out = 100)
predict_fits(fits, t, group = 1)

# Double-arm
fits <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  dist = c("exponential", "weibull")
)
t <- seq(from = 0, to = 1000, length.out = 100)
predict_fits(fits, t, group = 1)
predict_fits(fits, t, group = 2)
}