Launch Example Survival Analysis Script using the easy_adtte Data Set

This function launches an example script for starting survival analysis using the easysurv package. The script uses simulated phase III breast cancer trial data available from the ggsurvfit package. The code is inspired by usethis::use_template() but modified to work outside the context of an .RProj or package.

Usage

quick_start3(output_file_name = NULL)

Arguments

output_file_name

Optional. A file name to use for the script. Defaults to "easysurv_start.R" within a helper function.

Value

A new R script file with example code.

Examples

quick_start3()
#> ℹ easysurv template: Attempting to write a new .R file to a temporary directory.
#> ℹ Leaving /tmp/Rtmp1KlA1Y/easysurv_start.R unchanged.
#> ☐ Edit /tmp/Rtmp1KlA1Y/easysurv_start.R.
#> ℹ Remember to save the file to a permanent location if you wish to keep it.