Adds a custom theme for the package to a ggplot2 plot

The custom theme is based on theme_ipsum_rc from the hrbrthemes package. It uses the Roboto Condensed font, which is downloaded via sysfonts::font_add_google("Roboto Condensed") during package startup.

Usage

theme_easysurv(
  base_family = "Roboto Condensed",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 16,
  plot_title_face = "bold",
  plot_title_margin = 8,
  subtitle_family = base_family,
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 5,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 11,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(5, 5, 5, 5),
  panel_spacing = grid::unit(2, "lines"),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "grey30",
  axis = TRUE,
  ticks = FALSE
)

Arguments

base_family

Base font family for text elements in the plot (default is "Roboto Condensed").

base_size

Base font size for text elements in the plot (default is 11.5).

plot_title_family

Font family for plot titles.

plot_title_size

Font size for plot titles.

plot_title_face

Font face for plot titles (e.g., "plain," "bold," "italic," or combinations like "bold|italic") (default is "bold").

plot_title_margin

Margin between plot title and plot area (in points) (default is 8).

subtitle_family

Font family for subtitles.

subtitle_size

Font size for subtitles.

subtitle_face

Font face for plot titles (e.g., "plain," "bold," "italic," or combinations like "bold|italic") (default is "plain").

subtitle_margin

Margin between subtitle and plot area (in points) (default is 5).

strip_text_family

Font family for strip text (e.g., facet labels).

strip_text_size

Font size for strip text (default is 12).

strip_text_face

Font face for strip text (e.g., "plain," "bold," "italic," or combinations like "bold|italic") (default is "plain").

caption_family

Font family for plot captions.

caption_size

Font size for plot captions (default is 9).

caption_face

Font face for plot captions (e.g., "plain," "bold," "italic," or combinations like "bold|italic") (default is "plain").

caption_margin

Margin between captions and plot area (in points) (default is 10).

axis_text_size

Font size for axis labels (default is base_size; 11.5).

axis_title_family

Font family for axis titles.

axis_title_size

Font size for the axis titles (default is 11).

axis_title_face

Font face for axis titles (e.g., "plain," "bold," "italic," or combinations like "bold|italic") (default is "plain").

axis_title_just

Justification for the axis titles. It can be set to: "bl" (bottom left), "tl" (top left), "bc" (bottom center), "tc" (top center), "br" (bottom right), "tr" (top right) (default is "rt").

plot_margin

Margin around the entire plot area. Use the margin function (default is ggplot2::margin(5, 5, 5, 5)).

panel_spacing

Spacing between panels in a facetted plot. Use the unit function (default is grid::unit(2, "lines")).

grid_col

Color for grid lines (default is "#cccccc").

grid

Logical indicating whether to display grid lines.

axis_col

Color for axis lines (default is "grey30").

axis

Logical indicating whether to display axis lines.

ticks

Logical indicating whether to display ticks on the axis.

Value

An object of class theme that represents the custom theme.

Examples

if (FALSE) {
require("survival")
fit <- survfit(Surv(time, status) ~ as.factor(sex), data = lung)
ggsurvplot(fit, data = lung, ggtheme = theme_easysurv())
}